PS-2957 Document the release process
In attempting to make a release for the `Currency` changes,
it turned out that it wasn't clear how we make releases.

To help contributors, we want to make this a bit more visible.
So, we codify what the current process is.

This is a pretty standard process,
but we haven't explictly written it down anywhere.
As things grow and change, we can grow and change this documentation.