Dedent `where` clause for compatibility with 0.13.x
In 0.12.x,
the parser recognizes this clause as applying to the whole value.
In 0.13.x,
the parser recognizes this clause as applying to the `otherwise` clause.

The simplest thing is to dedent the `where` clause so it applyes
to the whole value regardless of which version of the compiler is used.
There is no semantic difference,
but this makes the codebase ready for 0.13.x.