Add support for deploying to k8s cluster
This is more or less copy-pasted from the Lynx repository. A common workflow for
some of us it to work on a remote server, where the best way to get browser
feedback is to deploy to a k8s cluster managed by helm. Notably we don't really
on outside configuration files like in our private repos, since this is a public
repo, it should be able to be deployed independently (assuming a compatible
deployment environment).

These configs could probably be cleaned up plenty, but wanted to just get this
working and move on to the actual task to be worked on.