More UI Guide fixes
Mainly was just trying out the UI guide as a test grounds to see if the UI
controls were working as expected. Ran into a few issues but it was only dealing
with the way the guide was implementing the controls, not the controls
themselves. I think this ocelot is halogen 5 ready now