Fix Header CSS
There was a small layout bug where we were giving the
`Ocelot.Block.Header.stickyHeader` block a height of `h-20`, but in usage, we
throw `Ocelot.Block.Header.header` inside which has a height of `h-24`. This
caused a discrepancy between the visual height of the header, and the height it
took up in the layout (`stickyHeader` gets fixed positioning, so the inner and
outer height have to match up so that content doesn't show up underneath the
block).