Fix `.clearfix` class
Seems Tailwind ships its own `.clearfix` class, but it doesn't work in all
situations (it only includes a rule for `.clearfix::after`, but we need the same
rule for `.clearfix::before` as well). We are using an older version of
Tailwind, it's possible this issue is fixed in a more recent version, but it's a
lot easier to fix the class ourselves than to upgrade a major version.