Fix bundling
When we moved to Spago, we had to update the way we were building and bundling
our UI Guide application, because the tools we were using were no longer
available. In the process we didn't realize that our tooling no longer bundled
for the browser automatically. We install `browserify` to fill that gap, and add
the extra step to our build. Since our build has now grown to several steps, we
break each out to their own command to make things easier to read, and so we can
run each individually, if desired.