FBCM-3571 Allow disabling Typeaheads dynamically
We want to allow Typeaheads to be enabled/disabled on the fly.

The way things have worked before, we would take the properties we
wanted to set on the Typeahead and apply them. The thing about those
props is that they are static. That method works, and has worked for us
for a long time.

What we found out, was that we wanted to have a Typeahead be enabled or
disabled on the fly. The approach passing in the properties doesn't work
because the properties don't change over the lifecycle of the component.

The way we addressed this took some work to implement. First, we
extended the state to know whether it's disabled or not. Then, we
extended the query to allow changing the disabled value on the state.
Finally, for backwards compatibility we made all of the helpers default
to whatever the properties were passed in. The backwards compatibility
should ensure that anything that was passing in static disabled
properties should continue to behave the same way.

Co-authored-by: David <dzuch@citizennet.com>