Add New Social Platform Icons
We have a host of new Platforms we wish to support in the UI, which means we
need a fresh set of icons. We've received a new icon set from the Design team,
and do the work here of adding them as icon fonts and making new blocks for them
in `Ocelot.Blocks.Icon`.

There are some steps involved in turning the SVG's into a font that's outside of
the workflow of this repo, but documentation for that process will be included
in a follow-up commit.