Update UI Guide to showcase new icons
Now that we have ten social icons, it makes sense to distinguish them from the
utility icons in the UI Guide, so we split the Icons page into two sections
representing each.

We also update our Tailwinds config with colors for the new social platforms.
This allows our icons to pop more, and helps with the mental association.