Vjp/fix modal scroll (#132)
* Fixes the modal screen overflow issue

The one where if the modal doesn't fit in the screen, the user can't do
anything about it (except zoom out in the browser).

This fixes it by moving the main modal window into the top level div
container, and adding an `overflow-y-auto` on that container to add in
scrollbars. We also get rid of the dark blanket div which used to rcv
click events to exit the modal. PM said that was actually undesired
bahavior

* Bump ocelot version to 0.23.1

* pr fixes

Bump ocelot to version 0.24.0