trigger date/time search on blur
We postpone parsing and selecting date/time through text input until user
presses Enter for confirmation which is reasonable. But this might lead to
inconsistent presentation of state between the input field and the calendar.
For example, when user clears the input field and unfocuses it, the selection is
expected to be cleared as well, but given current implementation the previous
selection persists.

To fix this issue, we force a search when the input field is unfocused just like
pressing Enter.