Use `Makefile` for building `css` directory
We want to decrease the amount of friction required for working on this
library. Having to remember a variety of commands necessary to get
started and to build the project is one source of friction we can
address. We start with moving the `css` directory, which is pretty
straightforward.

We also update the `package.json` scripts to use our new `Makefile`
rules. We don't want to maintain two different build systems, but we
at least keep around the commands in order to not disrupt existing
workflows. We can think about getting rid of them altogether at a later
time.