Use `Makefile` for build process
We finish the work we started in the last commit to move to `make` for
our builds. This will make it so developers don't have to worry about
remembering which commands to run and in which order, in order to get
setup and build their changes. That can all be taken care of by just
running `make`.

In addition to adding a `Makefile` to the root of our project, we also
update `css/Makefile` so that it doesn't conflict with the root one.

These updates also served to highlight a few issues with our setup. It
doesn't make much sense that the `css` directory has its own
`package.json` and `Makefile`, especially when `ui-guide` doesn't.
That's a bit backwards considering that building the `css` directory
outputs artifacts that we version control, whereas building the
`ui-guide` directory does not. We'll save that work for another time
though.

Again, we update the scripts in our `package.json` file to defer to
`make` instead of maintaining a separate set of build instructions. But
we go ahead and get rid of a few that were never really targets in and
of themselves, but instead steps to the actual target of `build-ui`.

There are a couple that didn't map perfectly, such as the `postinstall`
and `install-css` scripts. We should probably just get rid of them, but
we can leave that for later cleanup.

Lastly, we didn't port over the `debug` script, which uses `parcel`
instead of `browserify` to bundle `ui-guide`, because we didn't want to
do too much at once, and this isn't a target the author (davezuch) has
yet used.