Update CI to use `make`
We bring our CI config up to date with our build system. While we're at
it, we add in a check to make sure our build artifacts that we version
control aren't out of sync. I.e., that we haven't made changes to the
CSS but forgotten to build those changes and check them in.

The command was copied from:
https://circleci.com/developer/orbs/orb/niteo/check-untracked-changes

This may turn out to be more annoying than helpful, if somehow the CSS
builds aren't deterministic. If that turns out to be the case, we should
remove this verification step.