Add lockfile to version control
Adding the lockfiles to `.gitignore` sort of defeats the purpose. We
remove `yarn.lock` from the blacklist so that we can check it in. We
remove `package-lock.json` because it shouldn't exist at all. We use
`yarn` as our package manager, if we're producing `package-lock.json`
files, we're doing something wrong, and ignoring the file hides that
fact.

Sure enough, removing `package-lock.json` from the `.gitignore` showed
that we had a script that used `npm` instead of `yarn` for installing
dependencies. Instead of fixing the offending script, we just get rid of
the `package.json` scripts that don't defer to `make`, as we don't want
to maintain multiple build paths.