Move Typeahead modules
When we started this library, we thought it would be useful to namespace
modules by whether they export a Component, or Blocks. In practice, this
partitioning hasn't been so useful. It results in longer module names,
forces a segregation of Blocks and Components, and doesn't save the
developer from having to look at a module to see its contents when
working with either.

We would like to abandon the hierarchy altogether, but we don't want to
do it all at once, as that would be too big a change. We take a
piecemeal approach, starting with the Typeahead module.