Slider.Render.pixelToPercent
When position of MouseEvent is measured on page, the coordinate values (pageX,
pageY) returned from DOM API is in pixel. But within SVG we rely on user space
unit (defined in viewBox), and percentage when talking about position relative
to the starting point of the track in proportion to the width (in units) of the
track.

This function computes the frame width in unit, a scaling factor from pixel
to unit, normalize length measurement in pixel on the track to percent.