Consolidate Typeahead modules
Having the Typeahead modules spread out into three components seemed
like a good way to keep them organized in theory, but in practice it
resulted in more confusion than clarity. Maintaining separate modules
meant having to worry about dependency cycles, having to figure out the
most appropriate module to make a change or addition to, confusion about
which module to use, or having to search through multiple modules to
find a specific type or value.

We reverse course and consolidate our Typeahead modules into one. This
change appears significant — and indeed the amount of code moved around
is — but no functionality has changed. We even avoided cleaning up the
imports to be fully qualified in order to change as little as we could.
All that's changed is the location of the Typeahead types and values. We
also took care to export all the types and values that Wildcat and OMS
really on, so as not to accidentally break downstream codebases.