FBCM-4107 Test functionality we want to change
We want to make some changes to how typeahead results are filtered,
before we do that it's a good idea to write some tests. This will help
to make sure the changes we make don't break anything, as well as
improve our confidence about the behavior of the changes we introduce.