FBCM-4107 Rename `runFilter` operation
There are two filter operations that we perform on search results.
First, we allow filtering on the raw results, which is useful for
removing currently selected values. Second, we allow filtering on the
Fuzzy results, which has is useful for filtering out poor matches.

Right now only the former is configurable, but we'd like to configure
the latter as well. We rename the one that is configurable from
`runFilter` to `runFilterItems` to make room for another option.