FBCM-4107 Make Typeahead's Fuzzy result filtering configurable
We move the filtering of the `Data.Fuzzy.Fuzzy _` results from being ad
internal implementation, to one that's configurable. Most consumers
would rather not have to worry about this filtering though, so we
provide the filter function we were using both as the default behavior
of our convenience components, as well as a function that can be
imported and used in a custom component.

The tests continue to pass so we know we didn't mess anything up.