FBCM-4107 Move sorting to filter function
In addition to being able to control how results are filtered, we also
want to allow the user to control how results are sorted. Or more
concretely, in our specific use-case, we want to opt out of both and
just rely on the endpoint we're pulling results from to do both.

We could add another operation that the user can pass in, but really,
there's no reason to just move the sorting into the filter function that
the user supplies. At least, if there is one, it isn't apparent. It's
something we can easily add in later though.

We add a test verifying the expected behavior, so we know it works.