FBCM-4107 Provide set of components that don't filter or sort results
Now that filtering and sorting is configurable, we provide a pair of
convenience components configured not to sort or filter the search
results. We also update our UI Guide to use this new setting for the
example User typeaheads, as the endpoint they hit does that for us.

With our tests passing and the UI Guide examples working as expected, we
can be satisfied that we've reached our desired outcome.