Add support for dynamically disabling Dropdowns
In order to disable a Dropdown in the past, we had to set the `disabled`
HTML attribute on the dropdown element and rely on the browser disabling
the element's behavior. The problem with this approach was that the
state setting the disabled property was on the parent, not the Dropdown
component itself, and when the parent's state changed, that change
wasn't progagated to the Dropdown.

We update the Dropdown so that disabled state is a first-class citizen
and allow querying the component to update the disabled state. After
updating the UI Guide our dynamic toggling now works, so we know we're
done.