FBCM-4154 Don't guess time when explicitly selected
Previously, any blur event on the time picker input was treated as an
attempt to search for a time. This got us into trouble when we
explicitly selected a time and
`Ocelot.Components.TimePicker.Utils.guessTime` guessed incorrectly.

To avoid this unnecessary guess, we clear out the existing selection
when a value is typed into the input. Then when there is a blur event,
we only search if there's no selection.

This solves the immediate problem described in
https://citizennet.atlassian.net/browse/FBCM-4154 but doesn't work if
the user types anything into the input in the range of 12:00 AM to
12:59 AM.