FBCM-4154 Fix meridiem selection
Previously, if the user didn't type a "P" or a "p" after a number in the
time picker input, the format "HHmm" was assumed (24-hour time). This
meant that if the user explicity typed out "12:45 AM", the value
"12:45 PM" would result since using "HHmm", "1245" is equivalent to
"12:45 PM". In order to specify "12:45 AM" one would have to write "045"
or "0:45".

We fix this by adding the possibility that the user intended to specify
"AM" and the format should be "hhmm a" (12-hour time). We only fall back
to 24-hour time if "AM" or "PM" cannot be guessed.