fix Typeahead border glitch
When the container (inputGroup) has "flex" (display: flex), all items will have the
same height meaning items of less content will be assigned the same height as
the one with the most content. But when the container also has
"items-center" (align-items: center) or any item has
"self-center" (align-self: center), items' heights are no longer regulated by
the container, which cause the glitch that the boundary of addonLeft (search
box) disconnects with the boundary of addonRight (button) when the content of
search box cannot be fit into a single line.

We fix it by moving the "items-center" class name from container to each addon
so that all addons' height are the same and their contents are centered inside.