update Interval by Receive action
The `interval` state is managed by parent so it's synchronized by a receive
action.

TimePicker is a Select composite component wrapped by a transparent container
component for cleaner interface.
To implement Input receiver properly, we need PassingReceive action registered
on the container component to pass down new Interval, and a Receive action
registered on the embedded component to handle actual render state update.

When the old Time selection no longer falls within the new Interval, we clear
the selection and raise SelectionChanged output.