persist Aligned in State
Notice a bug on date selection: dates from previous month and next month are
available to select, but once selected the year and month are not switched
accordingly. Aligned is currently an intermediate result for building up
calendarItems :: Array CalendarItem for rendering so it's not persisted. But in
order to fix the bug, we need to cache Aligned for determining the selected date
is whether previous, this, or next month.