give user full control over text rendering
We used to take a default label String and a text render function for a single
selection with `Maybe` in `renderToolbarSearchDropdown`. Now we want to support
multiple selections with `Array`, the presentation of text as a summary of
selections could vary depending on different situations, e.g. a comma separated
list of all selections, or the count of selections, etc. so we let users to
provide the render function instead of making a decision for them.