fix parcel errors and add `.parcel-cache` to `.gitignore`
We need to add `type="module"` attribute for parcel v2 to recognize source with
module related syntax like `import` and `require`.

`affjax` has a `require` statement for `xhr2` which is a `node` dependency so we
need to set shim "alias" in `package.json` file according to the doc

https://parceljs.org/features/dependency-resolution/#shim-aliases

We didn't have to do this for parcel v1 so I guess there's a change in the
algorithm of package resolution to improve speed.