Expand MultiInput inner input field to fill the whole container
Currently the MultiInput component has an inner auto-sized input field
whose initial width corresponds to that of the placeholder text. This
makes is difficult for users to interact with the component, because in
order to start typing in, they must necessarily click on the placeholder
text, as clicking on anywhere else in the container wouldn't hit the
inner input field.