Rename Button module
We want to drop `Block` from our module namespaces, as it's not useful
information, and makes usage with fully qualified imports more verbose.

Our plan is to migrate modules as we work on them, so we don't have to
do too much at once. Since we're about to make changes to
`Ocelot.Block.Button`, first we rename it to `Ocelot.Button`.