Sort Button definitions
The definitions in this module have ended up a bit disorganized over
time. We sort them alphabetically in an effort to make the module easier
to navigate, and in hopes that it'll help the module to stay organized.

The diff may be hard to read, but the only change is that definitions
were moved around.