FBCM-5076 Add pill blocks
We add variants style buttons to `Button` which we call "pills". These
are intended for button-like elements that require a more "flat"
appearance.

We depart a bit from the traditional "block" pattern here, which has
proved difficult to scale, when there are a multiplicity of variants.

Instead, we use `Option` to provide a convenient API for taking
advantage of the various configurations we offer. We won't know for sure
without enough usage, but this does seem a major improvement over the
current "block" pattern, and we ought to consider adopting it more
broadly.