FBCM-5076 Demonstrate pill buttons in UI Guide
With the addition of pill buttons, we demonstrate some of the different
variations that are possible from the UI Guide. Worth noting is that
while with the old "block" pattern we were forced to use `<button>`
elements, with the new pattern we not only offer a `<button>` and an
`<a>` option, but we could actually use any HTML element we want.